Shift Table analysis to ResultTable Class
